app/requirements.txt
django-cors-headers==3.7.0
django==2.2.10
djangorestframework==3.12.2
gunicorn==20.0.4
psycopg2-binary==2.8.6
